+++ examples/deno/javascript/fixedExample.js
@@ -2,7 +2,7 @@ import {
   availableParallelism,
   FixedThreadPool,
   PoolEvents,
-} from 'jsr:@poolifier/poolifier-web-worker@^0.3.16'
+} from 'jsr:@poolifier/poolifier-web-worker@^0.3.17'
 
 const pool = new FixedThreadPool(
   availableParallelism(),