+++ examples/deno/javascript/multiFunctionExample.js
@@ -1,7 +1,7 @@
 import {
   availableParallelism,
   FixedThreadPool,
-} from 'jsr:@poolifier/poolifier-web-worker@^0.3.16'
+} from 'jsr:@poolifier/poolifier-web-worker@^0.3.17'
 
 const pool = new FixedThreadPool(
   availableParallelism(),