+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 
 ### Fixed
 
-- Ensure settings worker choice strategy options at runtime are propagated to
+- Ensure worker choice strategy options changes at runtime are propagated to
   poolifier workers.
 - Do not reset worker node usage statistics at worker choice strategy runtime
   change.